Update invocation of setup script and fix typo in install-vulkan.ps1
Updated the run command in the Ubuntu Github Workflow to properly invoke the setup script using 'bash'. This ensures the setup script works as expected in the Linux environment.

Also, fixed a typo in Scripts/install-vulkan.ps1 where 'Containd' was replaced with 'Contains' for proper checking of the Vulkan SDK version. This prevents false negative messages about incorrect Vulkan version.