Refactor install-vulkan scripts for clarity and correctness
Removed debug/unnecessary print statements and tidied up conditional checks in install-vulkan.sh and install-vulkan.ps1 scripts for improved clarity and ease of debugging. Also added informative print message in install-vulkan.sh when Vulkan SDK is found via environment variable, providing better feedback to user. Among the changes was the removal of an inconsistency in the Test-Path which was not being correctly checked. The scripts are now more streamlined and easier to understand.