Refine build scripts and remove unnecessary flags
Modified the Ubuntu GitHub workflow and the install-vulkan.ps1 script to refine the build process. Unnecessary vendored flags for SDL3MIXER, SDL3IMAGE, and SDL3TTF were removed from CMakeLists.txt.

The refinement of scripts ensures a smoother build process and the removal of unnecessary flags simplifies building processes. In the install-vulkan.ps1 script, Windows-specific '.exe' extension was added to the vulkaninfo command for accurate version detection on Windows platforms. The libopus-dev installation command was removed from the Ubuntu workflow as it is no longer necessary.