Add cstddef to MemoryManagerCore header
A cstddef library was included in the MemoryManagerCore.h header file as it provides many typical features of C programming, such as NULL and size_t types, that can be necessary for the functions and operations within this file. This addition may increase the versatility and ease the manipulation of memory-related operations.