Add GreenHills support and console link option
This commit adds GreenHills compiler support to the build presets. All "NinjaMulti" presets were replaced by the "GreenHill" preset for Linux configurations. This change was necessary for better compatibility in different development environments.

In addition, a link option for the console subsystem was introduced in the Tests cmake file. This ensures that the executed program can output to a console window, which is useful for debugging and observing program output during development.